fixed Photon managers and configuration
