+++ Assets/MirageXR/Networking/Prefabs/ConnectionManager.prefab
@@ -9,9 +9,10 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 4325252047169387943}
-  - component: {fileID: 8430906712691975549}
   - component: {fileID: 4994710183780181035}
   - component: {fileID: 5129707200660208892}
+  - component: {fileID: 7466890686127646672}
+  - component: {fileID: 5783437905079510518}
   m_Layer: 0
   m_Name: ConnectionManager
   m_TagString: Untagged
@@ -35,32 +36,6 @@ Transform:
   - {fileID: 4661928101840865624}
   m_Father: {fileID: 0}
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!114 &8430906712691975549
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 2228691272098697712}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: c3b2dfec39f015045b1dbe3b2d68d390, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  gameMode: 2
-  roomName: DefaultRoom
-  connectOnStart: 1
-  playerCount: 0
-  connectionCriterias: 1
-  additionalSessionProperties: []
-  runner: {fileID: 0}
-  userPrefab: {fileID: 2153310105290136259, guid: c911395726db65a4eb527260054bb7d3,
-    type: 3}
-  loadSceneMode: 1
-  onWillConnect:
-    m_PersistentCalls:
-      m_Calls: []
-  actualSessionProperties: []
 --- !u!114 &4994710183780181035
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -114,6 +89,44 @@ MonoBehaviour:
   AutoConnectAndJoin: 1
   UseFusionAppSettings: 1
   UseFusionAuthValues: 1
+--- !u!114 &7466890686127646672
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 2228691272098697712}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: c428ebe59e7d14f9c97ccf67373eb36a, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  gameMode: 2
+  roomName: MainRoom
+  connectOnStart: 1
+  playerCount: 0
+  connectionCriterias: 1
+  additionalSessionProperties: []
+  runner: {fileID: 0}
+  userPrefab: {fileID: 2153310105290136259, guid: c911395726db65a4eb527260054bb7d3,
+    type: 3}
+  loadSceneMode: 1
+  onWillConnect:
+    m_PersistentCalls:
+      m_Calls: []
+  actualSessionProperties: []
+--- !u!114 &5783437905079510518
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 2228691272098697712}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: faaabb0b8c1057e43aa88bd2b5d13ae8, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!1 &2409919775912430426
 GameObject:
   m_ObjectHideFlags: 0
