+++ Assets/MirageXR/Networking/Scripts/PhotonHardwareManager.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 722d4e293dfff3147bdd3a81c0951f4e
+guid: faaabb0b8c1057e43aa88bd2b5d13ae8
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
