+++ Assets/MirageXR/Networking/Scripts/PhotonNetworkRig.cs.meta
