+++ Assets/Resources/VuforiaConfiguration.asset
@@ -24,7 +24,7 @@ MonoBehaviour:
     shareRecordingsInITunes: 0
     logLevel: 0
     version: 10.25.4
-    eulaAcceptedVersions: '{"Values":["0.0","8.3","8.1","7.2","7.5","8.0","8.6","8.5","9.3","9.4","9.6","9.7","9.5","9.8","10.6","10.11","10.10","10.15","10.19","10.25"]}'
+    eulaAcceptedVersions: '{"Values":["0.0","8.3","8.1","7.2","7.5","8.0","8.6","8.5","9.3","9.4","9.6","9.7","9.5","9.8","10.6","10.11","10.10","10.15","10.19","10.25","7.1"]}'
   database:
     disableModelExtraction: 0
   shaders:
