--- Assets/XRProject.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 1054a415cf8e3a3448fa9db0715cb7ee
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
