--- Assets/XRProject/Prefabs.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: f3643975b8c866d4990abbe50c316520
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
