--- Assets/XRProject/Prefabs/MRTKNetworkRigVariant.prefab.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: bf62f93f456c1ed40a08f5b514643ba3
-labels:
-- FusionPrefab
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
